test: split the four suites that outgrew one concept, and name the phases of the four longest scenarios
The last size findings, all in integration tests. Assertions are
preserved 1:1 — the test-name sets are identical before and after (917
in compose/integration, 1344 across compose + overlay), and every
assertion count matches per file: backfill 54, captureverdict 115,
jobs_overlay 85, teardown 56.

Long scenarios keep their sequence and gain names for their phases, so
the test body now reads as the scenario: TestBackfillWire 277→12 (its
nineteen t.Run phases become named helpers, the four op closures become
backfillWireEnv methods), TestOverlayReadAndSyncEndToEnd 269→20,
TestDisconnectPurgesTheMirrorTombstonesAndRetainsTheConnectionAudit
206→22 (one helper per claim, so purge, tombstone and audit-retention
still fail separately), and the AC-OV-2 read-subset case 191→125. Every
asserting helper calls t.Helper() so a failure still names the caller's
line.

Files split on subject seams, never at a line count:
captureverdict loses the human-review lane and the noise-reach lane;
jobs_overlay loses the disconnect-fence proofs; org360_suggestions
becomes what the card computes, the dismissal lifecycle, and
signal-derived advice; overlay_acceptance loses AC-OV-1 (the only
stdlib-only, database-free criterion — it is proved by the import
graph's shape) and the two budget-metered criteria. Each new file
carries the same //go:build integration tag, and no shared helper is
duplicated.

Groundwork for craft static --strict.
